Refactor vault_wallet.go, moving Hashicorp specific code to new file
